# public/api/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "POS API",
        "description": "REST API untuk Sistem Point of Sale (POS)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8303/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/penjualans": {
            "get": {
                "tags": [
                    "Penjualans"
                ],
                "summary": "Dapatkan semua penjualan (filter & pagination)",
                "operationId": "getPenjualans",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-31"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pagination"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Penjualans"
                ],
                "summary": "Buat transaksi penjualan",
                "operationId": "createPenjualan",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PenjualanCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Penjualan berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/penjualans/{penjualan}": {
            "get": {
                "tags": [
                    "Penjualans"
                ],
                "summary": "Dapatkan detail penjualan",
                "operationId": "getPenjualanDetail",
                "parameters": [
                    {
                        "name": "penjualan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "INV-20250101-0001"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/penjualans/summary": {
            "get": {
                "tags": [
                    "Penjualans"
                ],
                "summary": "Dapatkan rangkuman penjualan (summary)",
                "operationId": "getPenjualanSummary",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Summary data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "periode": {
                                            "properties": {
                                                "start_date": {
                                                    "type": "string"
                                                },
                                                "end_date": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "total_penjualan": {
                                            "type": "number"
                                        },
                                        "total_diskon": {
                                            "type": "number"
                                        },
                                        "total_pajak": {
                                            "type": "number"
                                        },
                                        "total_laba": {
                                            "type": "number"
                                        },
                                        "jumlah_transaksi": {
                                            "type": "integer"
                                        },
                                        "rata_rata_transaksi": {
                                            "type": "number"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ItemPenjualan": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Pagination"
                    },
                    {
                        "properties": {
                            "kode_barang": {
                                "type": "string",
                                "example": "BRG001"
                            },
                            "qty": {
                                "type": "integer",
                                "example": 2
                            },
                            "harga_satuan": {
                                "type": "number",
                                "format": "float",
                                "example": 15000
                            },
                            "jumlah": {
                                "type": "number",
                                "format": "float",
                                "example": 30000
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "PenjualanCreate": {
                "required": [
                    "id_nota",
                    "tgl",
                    "items"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Pagination"
                    },
                    {
                        "properties": {
                            "id_nota": {
                                "type": "string",
                                "example": "INV-20250101-0001"
                            },
                            "tgl": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-01 10:00:00"
                            },
                            "kode_pelanggan": {
                                "type": "string",
                                "example": "PLG001",
                                "nullable": true
                            },
                            "diskon": {
                                "type": "number",
                                "example": 0
                            },
                            "pajak": {
                                "type": "number",
                                "example": 0
                            },
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/ItemPenjualan"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ApiResponse": {
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Pagination": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "type": "string"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "last_page_url": {
                        "type": "string"
                    },
                    "next_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "prev_page_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}